requirements.txt
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
